requirements.txt
# Core dependencies
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
Authlib==1.6.0
cachetools==6.1.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
confluent-kafka==2.3.0
cryptography==45.0.5
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jsonschema==4.24.1
jsonschema-specifications==2025.4.1
packaging==25.0
pluggy==1.6.0
protobuf==3.20.3
pycparser==2.22
Pygments==2.19.2
pytest==8.4.1
python-dotenv==1.1.1
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
sniffio==1.3.1
typing_extensions==4.14.1
typing-inspection==0.4.1
urllib3==2.5.0

# FastAPI stack
fastapi==0.116.1
starlette==0.47.1
uvicorn==0.35.0

# Pydantic
pydantic==2.11.7
pydantic_core==2.33.2
pydantic-settings==2.10.1

# Kafka ecosystem
fastavro==1.11.1
python-schema-registry-client==2.6.1

# Protobuf & gRPC
grpcio==1.73.1
googleapis-common-protos==1.70.0

# Legacy 
avro==1.12.0

# Дополнительные пакеты 
aiofiles==24.1.0
asyncpg==0.29.0
greenlet==3.2.3
SQLAlchemy==2.0.25







